+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+#### --- Version 0.6.13
+This version update:
+* Added new element `divider`
+* Fixed style of `alert`, `button`, `switch` (no need to required TailwindCss after install)
+* Changed attribute from `show` to `v-model` of `Alert` component and fixed auto-dismiss update modelValue
+* Fixed `button` and `input` inherit Attributes
+* Added Readme (Vietnamese language)
+
 #### --- Version 0.5.0
 This version update:
 * Added new element `switch`